+++ src/main/java/Zeus/API/ZEUS/Dto/DadosListagemRacao.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Date;
 
@@ -15,7 +16,7 @@ public record DadosListagemRacao(
         @NotNull
         int kqQuantidade,
         @NotNull
-        int valorPago,
+        double valorPago,
         @NotNull
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
         LocalDate dataCompra
